+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * [UIPQB-6](https://issues.folio.org/browse/UIPQB-6) Array property type - action-value component.
 * [UIPQB-14](https://issues.folio.org/browse/UIPQB-14) Build query string
 * [UIPQB-17](https://issues.folio.org/browse/UIPQB-17) Test query
+* [UIPQB-27](https://issues.folio.org/browse/UIPQB-27) Title: Extend QueryBuilder plugin with 'saveBtnLabel' Prop